Move error_prop, ratios, and temp_metallicity_calc into analysis sub-package (iss #39)
