+++ Metallicity_Stack_Commons/analysis/error_prop.py
@@ -5,7 +5,7 @@ from . import line_name
 from astropy.io import ascii as asc
 import numpy as np
 
-from .column_names import filename_dict, temp_metal_names0, npz_filename_dict
+from ..column_names import filename_dict, temp_metal_names0, npz_filename_dict
 from .ratios import error_prop_flux_ratios
 from .temp_metallicity_calc import temp_calculation, metallicity_calculation
